fixed bug when resizing hashmap and hashset, and also changed hashset to support size_t

+++ HashSet.h
@@ -2,6 +2,7 @@
 #define HASH_SET_H
 
 #include <stdlib.h>
+#include <stdint.h>
 #include "LinkedList.h"
 
 typedef struct HashSet {
@@ -11,8 +12,8 @@ typedef struct HashSet {
 } HashSet;
 
 void initHashSet(HashSet* set, size_t initialCapacity);
-void insertHashSet(HashSet* set, int key);
-int containsHashSet(HashSet* set, int key);
+void insertHashSet(HashSet* set, size_t key);
+int containsHashSet(HashSet* set, size_t key);
 void destroyHashSet(HashSet* set);
 
 #endif
